requirements.txt
arrow==0.13.1
blessed==1.15.0
certifi==2019.3.9
chardet==3.0.4
Django==1.11.11
django-ace==1.0.5
django-cors-headers==2.5.2
django-filter==2.1.0
django-macaddress==1.5.0
django-picklefield==2.0
django-q==1.0.0
django-recurrence==1.7.0
djangorestframework==3.9.2
get==2019.3.22
idna==2.8
Markdown==3.1
netaddr==0.7.19
post==2019.3.22
psycopg2==2.8.1
public==2019.3.22
python-dateutil==2.8.0
pytz==2019.1
query-string==2019.3.22
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
wcwidth==0.1.7
